Switch to only using sprites in Loenn, and add arrows
Also, I added a second room. It seems I'll have to add the room position
to node positions in the C# code, since entity positions are in level
coords, but node positions are in room coords for some reason??? Room
transitions otherwise seem to work well with teleporting though, which
is nice.