Inter-room teleports, fix crash on death.
Also reversed the direction of the arrow in Loenn so that it shows the
direction the "pixels" were moved in (ie it points at where the portal
is rendered, not away from it).

Also fixed portals and teleporters in rooms that aren't at 0,0, oops.
Wasn't using the offset parameter correctly in the constructor for both
entities.