The Recursion Update
Yay, complete rework of the portal rendering :D

Currently it uses the turning number of the polygon for culling portal
renders, but this is incorrect. I plan to replace it with a culling
technique based on angle intervals, which I have started to implement a
struct for in AngleInterval.cs. This struct will represent a contiguous
set of angles relative to the positive x axis (I think), originating
from the player. That way, a narrow open portal will only require
rendering a few portals inside it, if that.