Angle interval culling
Ugh, floating point precision bugs. I think most of them are gone, but
there still are cases with closed polygons that aren't getting properly
culled for some reason. I don't know why haha. In theory, the bounding
box culling should be limiting it, but the closed polygon in the nyoom
room still is causing unbounded recursion.

For now, I added a setting to enable / disable the recursion limits
defined in the map. These limits work to cut the unbounded recursion
very early, but ideally the culling should be doing this for the map
maker, instead of forcing the map maker to manually define the recursion
limits.