Add radius-based axis aligned polar coord culling
There already was angle based culling, so I had the idea that maybe
radius based culling would help somewhat with some of the overlapping
portal cases. Also has the fun effect of letting me sort the portals
based on average distance to the player, which together work well in
most cases. There are still spots in the maze where the culling still
fails however; this is caused by the furthest point of a portal at the
start that shouldn't be rendered being further from the player than the
closest point of the portal the player is looking through, causing the
radius based culling to fail. There are two potential solutions:
1. Implement arbitrary polygon intersection based culling, replacing all
   existing culling methods. This is the most general form of culling,
   however the implementation would be quite time consuming to write.
2. Use shaders to cover the remaining cases, using a fragment shader to
   implement proper near-surface clipping. This wouldn't be too bad,
   using a color channel to represent a depth from 0 to 1 where 0 is
   used for the vertices infinitely far away and 1 is used for the
   vertices in the level. Then fragments can be discarded based on this
   information. This would still be tricky to implement, mainly due to
   needing to learn how to write custom XNA shaders, but it is something
   that would be good to learn anyway, especially if I want to add
   borders to the open polygon portals.