Maze improvements
Also, just noticed a bug with how the angle culling works, specifically
if one goes to where it says meow in the map, and then go in the bottom
right corner and jump, the portal render covering the girders gets
culled too early. It's happening because we are in fact on the wrong
side of the line connecting the endpoints, however because of the part
of the render that juts out to the bottom left, it should still be
rendered. TBD how the angle culling should work to handle this case
correctly.